use CodeBlock from components
