+++ apps/source.coop/eslint.config.mjs
@@ -13,7 +13,7 @@ export default tseslint.config(
     files: ['**/*.{ts,tsx,js}'],
     languageOptions: {
       ecmaVersion: 2020,
-      globals: globals.node,
+      globals: [globals.node, globals.browser],
       parserOptions: {
         project: './tsconfig.eslint.json',
         tsconfigRootDir: import.meta.dirname,
