+++ packages/components/src/components/CodeBlock.tsx
@@ -1,8 +1,36 @@
-import React, { ComponentPropsWithoutRef, useRef } from 'react'
+import React, { ComponentPropsWithoutRef, useRef, useState } from 'react'
 import { toast } from 'react-hot-toast'
 import { Box, Card } from 'theme-ui'
 import { useCopyToClipboard } from 'usehooks-ts'
 import SVG from './SVG'
+
+export function ClipboardIcon() {
+  return (
+    <SVG
+      viewBox="0 0 16 16"
+      sx={{ fill: 'primary', '&:hover': { fill: 'secondary' } }}
+    >
+      <path d="M4 1.5H3a2 2 0 0 0-2 2V14a2 2 0 0 0 2 2h10a2 2 0 0 0 2-2V3.5a2 2 0 0 0-2-2h-1v1h1a1 1 0 0 1 1 1V14a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1V3.5a1 1 0 0 1 1-1h1z" />
+      <path d="M9.5 1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-3a.5.5 0 0 1-.5-.5v-1a.5.5 0 0 1 .5-.5zm-3-1A1.5 1.5 0 0 0 5 1.5v1A1.5 1.5 0 0 0 6.5 4h3A1.5 1.5 0 0 0 11 2.5v-1A1.5 1.5 0 0 0 9.5 0z" />
+    </SVG>
+  )
+}
+
+export function ClipboardCopiedIcon() {
+  return (
+    <SVG
+      viewBox="0 0 16 16"
+      sx={{ fill: 'primary', '&:hover': { fill: 'secondary' } }}
+    >
+      <path
+        fillRule="evenodd"
+        d="M10.854 7.146a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708 0l-1.5-1.5a.5.5 0 1 1 .708-.708L7.5 9.793l2.646-2.647a.5.5 0 0 1 .708 0"
+      />
+      <path d="M4 1.5H3a2 2 0 0 0-2 2V14a2 2 0 0 0 2 2h10a2 2 0 0 0 2-2V3.5a2 2 0 0 0-2-2h-1v1h1a1 1 0 0 1 1 1V14a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1V3.5a1 1 0 0 1 1-1h1z" />
+      <path d="M9.5 1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-3a.5.5 0 0 1-.5-.5v-1a.5.5 0 0 1 .5-.5zm-3-1A1.5 1.5 0 0 0 5 1.5v1A1.5 1.5 0 0 0 6.5 4h3A1.5 1.5 0 0 0 11 2.5v-1A1.5 1.5 0 0 0 9.5 0z" />
+    </SVG>
+  )
+}
 
 export function InlineCode({ children }: {children: React.ReactNode}) {
   return (
@@ -29,7 +57,7 @@ type CodeBlockProps = {
 
 export default function CodeBlock({ children, copyButton, ...props }: CodeBlockProps) {
   const copy = useCopyToClipboard()[1]
-
+  const [copied, setCopied] = useState(false)
   const buttonRef = useRef<HTMLElement | undefined>(undefined)
   return (
     <Card variant="code" {...props}>
@@ -47,18 +75,13 @@ export default function CodeBlock({ children, copyButton, ...props }: CodeBlockP
           backgroundColor: 'muted',
         }}
         onClick={() => {
+          setCopied(true)
           if (buttonRef.current) {
             copy(copyToClipboard(buttonRef.current)).catch((e: unknown) => { console.error(e) })
           }
         }}
       >
-        <SVG
-          viewBox="0 0 16 16"
-          sx={{ fill: 'primary', '&:hover': { fill: 'secondary' } }}
-        >
-          <path d="M4 1.5H3a2 2 0 0 0-2 2V14a2 2 0 0 0 2 2h10a2 2 0 0 0 2-2V3.5a2 2 0 0 0-2-2h-1v1h1a1 1 0 0 1 1 1V14a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1V3.5a1 1 0 0 1 1-1h1v-1z" />
-          <path d="M9.5 1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-3a.5.5 0 0 1-.5-.5v-1a.5.5 0 0 1 .5-.5h3zm-3-1A1.5 1.5 0 0 0 5 1.5v1A1.5 1.5 0 0 0 6.5 4h3A1.5 1.5 0 0 0 11 2.5v-1A1.5 1.5 0 0 0 9.5 0h-3z" />
-        </SVG>
+        {copied ? <ClipboardCopiedIcon /> : <ClipboardIcon />}
       </Box>
       }
       <Box ref={buttonRef}>{children}</Box>
