+++ packages/components/src/index.ts
@@ -2,7 +2,7 @@ export { default as AccountDropdown } from './components/AccountDropdown'
 export { default as BaseLayout } from './components/BaseLayout'
 export { default as Browser } from './components/Browser'
 export { default as Button } from './components/Button'
-export { default as CodeBlock } from './components/CodeBlock'
+export { default as CodeBlock, InlineCode } from './components/CodeBlock'
 export { default as Footer } from './components/Footer'
 export { default as Form } from './components/Form'
 export type { FormFieldState, FormResult } from './components/Form'
